requirements.txt
anthropic>=0.25.0
msal>=1.28.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
flask>=3.0.0
